% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inheritance.R
\name{Inheritance}
\alias{Inheritance}
\alias{G}
\alias{gtype_init}
\alias{gamete_frequency}
\alias{gamete_frequency_by_morph}
\title{Inheritance}
\usage{
G(r = 0)

gtype_init(gtype)

gamete_frequency(gtype, G)

gamete_frequency_by_morph(gtype, G)
}
\arguments{
\item{r}{numeric(1) recombination rate, between 0 (complete
linkage) and 1 (free recombination).}

\item{gtype}{named numeric() vector of genotype frequencies. The
vector must have at least one positive value; it cannot contain
NAs. The names of the vector correspond to the \code{Genotype}
column of the \code{genetics} data object; names cannot be
duplicated. If missing, genotypes are set to approximate
isoplethy under complete disassortative mating.}

\item{G}{numeric() matrix of genotype-to-gamete transition
probabilities, from \code{G()}.}
}
\value{
A vector of genotype frequencies (i.e., non-negative values
    summing to 1)

numeric() matrix of genotype x gamete expected frequencies

numeric() matrix of morph x gamete expected frequencies.
}
\description{
Inheritance
}
\section{Functions}{
\itemize{
\item \code{G}: Create a matrix summaring gametes produced
by each genotype, allowing for arbitrary recombination between
the S and M loci.

\item \code{gtype_init}: Create initial genotype frequencies in
standard form.

\item \code{gamete_frequency}: Calculate gamete freqencies from current
genotype frequencies and mode of inheritance.

\item \code{gamete_frequency_by_morph}: Calculate gamete frequencies produced by
each morph from current genotype frequencies and mode of
inheritance.
}}
\examples{
G(0)
G(1)
G(.1)

gtype_init()    # approximate isoplethy
gtype_init(sm/sm=100, sM/sm=70, sM/sM=20, Sm/sm=50)
}

