% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mating.R
\name{Mating}
\alias{Mating}
\alias{M}
\alias{morph_frequency}
\alias{mate}
\title{Mating}
\usage{
M()

morph_frequency(population)

mate(population, G = Tristyly::G(0), M = Tristyly::M())
}
\arguments{
\item{population}{A numeric() vector of genotype counts, in the
same form as returned by \code{gtype_init()}, but with
\code{sum(population)} the size of the population.}

\item{G}{See \code{G()}.}

\item{M}{See \code{M()}.}
}
\value{
A 3 x 3 numeric matrix of fertilities

named numeric(3) vector of morph frequencies

numeric() vector of sampled genotypes.
}
\description{
Mating
}
\section{Functions}{
\itemize{
\item \code{M}: Create a matrix describing dissasortative mating
and morph fertility. The i,jth entry in the matrix is the
contribution of pollen from morph j to the pollen pool of morph
i. \code{rowSums(M())} are morph-specific female
fertilities. Currently only pure dissassortative mating is
supported

\item \code{morph_frequency}: Summarize population morph frequencies.

\item \code{mate}: Mate individuals in \code{population} using the
inheritance model defined by \code{G} and mating patterns
summarized by \code{M}. Only pure disassortative mating and
equal morph-specific female fertility are currently supported.
}}
\examples{
M()
gtype <- gtype_init()
n <- setNames(as.integer(gtype * 1000), names(gtype))
morph_frequency(n)
n1 <- mate(n)
n1
morph_frequency(n1)
}

