% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plotting}
\alias{Plotting}
\alias{plot_iterations}
\title{Plotting}
\usage{
plot_iterations(iterations, ..., type = "b", add = FALSE)
}
\arguments{
\item{iterations}{An object returned by \code{\link{iterate}}.}

\item{...}{Additional graphical parameters used by \code{lines()}
to plot population trajectories.}

\item{type}{character(1) Type of plot to create; default plots
trajectories as points joined by lines.}

\item{add}{logical(1) Create a new plot (\code{add = FALSE},
default) or add a trajectory to an existing plot.}
}
\value{
None; invoked for side effect.
}
\description{
Plotting
}
\section{Functions}{
\itemize{
\item \code{plot_iterations}: Plot simulated changes in morph frequency.
}}
\examples{
plot_iterations(iterate(isoplethy(50), 1000), pch=20)
}

