% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inheritance.R
\name{Inheritance}
\alias{Inheritance}
\alias{G}
\alias{gtype_init}
\title{Inheritance}
\usage{
G(r)

gtype_init(gtype)
}
\arguments{
\item{r}{numeric(1) recombination rate, between 0 (complete
linkage) and 1 (free recombination).}

\item{gtype}{named numeric() vector of genotype frequencies. The
vector must have at least one positive value; it cannot contain
NAs. The names of the vector correspond to the \code{Genotype}
column of the \code{genetics} data object; names cannot be
duplicated. If missing, genotypes are set to approximate
isoplethy under complete disassortative mating.}
}
\value{
A vector of genotype frequencies (i.e., non-negative values
    summing to 1)
}
\description{
Inheritance
}
\section{Functions}{
\itemize{
\item \code{G}: Create a matrix summaring gametes produced
by each genotype, allowing for arbitrary recombination between
the S and M loci.

\item \code{gtype_init}: Create initial genotype frequencies in
standard form.
}}
\examples{
G(0)
G(1)
G(.1)

gtype_init()    # approximate isoplethy
gtype_init(sm_sm=100, sM_sm=70, sM_sM=20, Sm_sm=50)
}

