% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inheritance.R
\name{Inheritance}
\alias{Inheritance}
\alias{G}
\alias{isoplethy}
\alias{as_genotype}
\title{Inheritance}
\usage{
G(r = 0)

isoplethy(N)

as_genotype(gtype)
}
\arguments{
\item{r}{numeric(1) recombination rate, between 0 (complete
linkage) and 1 (free recombination).}

\item{N}{missing or integer(1). When missing, return genotype
frequencies in populations at approximate isoplethic
equilibirum. Otherwise, return a sample of size \code{n} drawn
from a poualation at approximate isoplethy.}

\item{gtype}{named numeric() vector of genotype frequencies. The
vector must have at least one positive value; it cannot contain
NAs. The names of the vector correspond to the \code{Genotype}
column of the \code{genetics} data object; names cannot be
duplicated. If missing, genotypes are set to approximate
isoplethy under complete disassortative mating.}
}
\value{
A vector of genotype frequencies (i.e., non-negative values
    summing to 1)
}
\description{
Inheritance
}
\section{Functions}{
\itemize{
\item \code{G}: Create a matrix summaring gametes produced
by each genotype, allowing for arbitrary recombination between
the S and M loci.

\item \code{isoplethy}: Create initial, approximately isoplethic,
genotype frequencies in standard form.

\item \code{as_genotype}: Create a vector of genotype frequencies in
standard form.
}}
\examples{
G(0)
G(1)
G(.1)
isoplethy()     # approximate isoplethy
isoplethy(100)  # sample from isoplethic population
}

