% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iteration.R
\name{Iteration}
\alias{Iteration}
\alias{iterate}
\alias{iterate_to_monomorphism}
\alias{iterate_to_dimorphism}
\title{Iteration}
\usage{
iterate(population, generations, G = Tristyly::G(), M = Tristyly::M())

iterate_to_monomorphism(population, times, G = Tristyly::G(),
  M = Tristyly::M(), progress.interval = 10)

iterate_to_dimorphism(population, times, G = Tristyly::G(),
  M = Tristyly::M(), progress.interval = 50)
}
\arguments{
\item{population}{See \code{\link{Mating}}.}

\item{generations}{the number of generations to iterate; the initial
population is generation 1.}

\item{G}{See \code{\link{G}()}.}

\item{M}{See \code{\link{M}()}.}

\item{times}{numeric(1) times to iterate the population to
monomorphism.}

\item{progress.interval}{numeric(1) report progress every \code{verbose}
iterations. Suppress output with \code{verbose = Inf}.}
}
\value{
a tibble (data.frame) summarizing morph frequencies in each
    generation, including the first (generation 1)
}
\description{
Iteration
}
\section{Functions}{
\itemize{
\item \code{iterate}: Iterate a single population for a fixed
number of generations.

\item \code{iterate_to_monomorphism}: Repeatedly iterate a population until
monomorphism.

\item \code{iterate_to_dimorphism}: Repeatedly iterate a population until loss of
one morph.
}}
\examples{
population <- rmultinom(1, 10, gtype_init())[,1]
iterate(population, 1000)
population_size <- 10
population <- rmultinom(1, population_size, gtype_init())[,1]
morph_frequency(population)
mono <- iterate_to_monomorphism(population, 30)
table(mono$Genotype)
plot(ecdf(mono$Generation), xlab="Generation")
}

