% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plotting}
\alias{Plotting}
\alias{plot_iterations}
\title{Plotting}
\usage{
plot_iterations(iterations, ..., pch, type = "b", add = FALSE)
}
\arguments{
\item{iterations}{An object returned by
\code{\link{iterate_stochastic}} or
\code{\link{iterate_deterministic}}.}

\item{...}{Additional graphical parameters used by \code{lines()}
to plot population trajectories.}

\item{pch}{Plotting symbol(s); see \code{points}. If missing, the
first point is a solid dot and subsequent points open circles.}

\item{type}{character(1) Type of plot to create; default plots
trajectories as points joined by lines.}

\item{add}{logical(1) Create a new plot (\code{add = FALSE},
default) or add a trajectory to an existing plot.}
}
\value{
None; invoked for side effect.
}
\description{
Plotting
}
\section{Functions}{
\itemize{
\item \code{plot_iterations}: Plot changes in morph frequency across
generations. The starting population is represented by a solid
dot.
}}
\examples{
plot_iterations(iterate_stochastic(isoplethy(50), 1000), pch=20)
}

